my-ai-saas/scripts/requirements.txt
# News Scraper Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10
APScheduler>=3.10.0

# Optional: Advanced scraping
newspaper3k>=0.2.8
selenium>=4.15.0  # For JavaScript-heavy sites
python-dateutil>=2.8.0

# Database (choose one)
# For SQLite (lightweight)
# sqlite3 (built into Python)

# For PostgreSQL
# psycopg2-binary>=2.9.0

# For MySQL
# PyMySQL>=1.1.0
